requirements.txt
mysql-connector-python==9.1.0
pandas
flask